Respond to all requests with 503 error page
Respond statically to requests for the CSS file, and respond with the index.html file for anything else.

Always respond with 503 status code, as the service is unavailable. We can't use `serveFile` for this because it forces 200 responses, but it's also doing a bunch of stuff we don't need anyway.

#minor